[projectfirma/#2155] Project Running Balance: use Larry's Special as source Adding totals footer.
