+++ Source/ProjectFirma.Web/Views/Shared/ProjectRunningBalance/ProjectRunningBalanceViewData.cs
@@ -19,10 +19,8 @@ Source code is available upon request via <support@sitkatech.com>.
 </license>
 -----------------------------------------------------------------------*/
 
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using ProjectFirmaModels.Models;
 
 namespace ProjectFirma.Web.Views.Shared.ProjectRunningBalance
 {
@@ -30,56 +28,9 @@ namespace ProjectFirma.Web.Views.Shared.ProjectRunningBalance
     {
         public List<ProjectRunningBalanceRecord> ProjectRunningBalanceRecords { get; set; }
 
-
         public ProjectRunningBalanceViewData(List<ProjectRunningBalanceRecord> projectRunningBalanceRecords)
         {
             ProjectRunningBalanceRecords = projectRunningBalanceRecords.OrderBy(x => x.Date).ToList();
         }
-    }
-
-    public class ProjectRunningBalanceRecord
-    {
-        /// <summary>
-        /// Posting Date from Obligation Item Budget/Invoice
-        /// or Date from ProjectedBudget
-        /// </summary>
-        public DateTime Date { get; set; }
-
-        /// <summary>
-        /// Dummy data - eventually coming from user entered data in UI
-        /// </summary>
-        public double ProjectedBudget { get; set; }
-        public double ObligationItemBudgetObligation { get; set; }
-        public double ObligationItemInvoiceDebit { get; set; }
-
-        /// <summary>
-        /// Constructor for building a Project Running Balance Record for a budget projection
-        /// </summary>
-        /// <param name="projectedBudget"></param>
-        /// <param name="dateOfProjectedBudget"></param>
-        public ProjectRunningBalanceRecord(double projectedBudget, DateTime dateOfProjectedBudget)
-        {
-            ProjectedBudget = projectedBudget;
-            Date = dateOfProjectedBudget;
-            ObligationItemBudgetObligation = 0;
-            ObligationItemInvoiceDebit = 0;
-        }
-
-        public ProjectRunningBalanceRecord(WbsElementObligationItemBudget obligationItemBudget)
-        {
-            ProjectedBudget = 0;
-            Date = obligationItemBudget.PostingDateKey ?? DateTime.MinValue;
-            ObligationItemBudgetObligation = obligationItemBudget.Obligation ?? 0;
-            ObligationItemInvoiceDebit = 0;
-        }
-
-        public ProjectRunningBalanceRecord(WbsElementObligationItemInvoice obligationItemInvoice)
-        {
-            ProjectedBudget = 0;
-            Date = obligationItemInvoice.PostingDateKey ?? DateTime.MinValue;
-            ObligationItemBudgetObligation = 0;
-            ObligationItemInvoiceDebit = obligationItemInvoice.DebitAmount ?? 0;
-        }
-
     }
 }
